代码优化

+++ easy-elasticsearch-client/src/main/java/com/easy/elastic/search/result/SearchPageResult.java
@@ -32,7 +32,7 @@ public class SearchPageResult<T> implements Serializable {
     /**
      * 页码总数
      */
-    private int totalPage;
+    private long totalPage;
     /**
      * 数据集合
      */
